backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
flask-cors==3.0.10
